requirements.txt
colorama==0.4.6
graphviz==0.20.3
PyYAML==6.0.1
